Refactor authentication module: update imports from interfaces to types, and introduce new type definitions for authentication and user entities to improve code organization and clarity.
